requirements.txt
aiosmtplib==5.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==5.0.0
billiard==4.2.4
blinker==1.9.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
currencyapicom==0.1.1
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
everapi==0.1.1
fastapi==0.128.0
freezegun==1.5.5
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
iso4217parse==0.6.2
kombu==5.6.2
loguru==0.7.3
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==1.7.1
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
redis==7.1.0
regex==2025.11.3
requests==2.32.5
ruff==0.15.0
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.2.14
